client/requirements.txt
fluentd-logger
